Build out initial ImageTemplate
Adding some structure around the generation of an image.

An image template can take in a hash of options (these are passed to an
ERB template) and then rendered by to_html.

Also thinking this can give us flexibility to validate/make changes to
the options when they come in by overriding the initializer.