Build Authorization helper to limit access to ShareMeow
We don't want just anyone generating images with ShareMeow.

So we'll require that each request is accompanied with a hex encoded HMAC digest.

On request, if the digest is correct. Then the image can be generated.
Otherwise, we will 401.

This key can be changed at anytime.

This method inspired by how Camo does authorization:
https://github.com/atmos/camo. Thx camo! <3