Refactor image templates. set allowed_options in template
Ok, I'm not totally happy with this. But is better than what I had
before.

Sinatra does this confusing thing where Params come in as an
indifferent access hash (but class name is hash). Aka, both symbols and
strings work as keys.

I want to use symbols because I'd rather not be freezing strings
everywhere.

When options come in now, it will check against the allowed list (of
strings). And then anything on that list will by symbolized.

I don't want to symbolize everything because then we'd be storing any
random param as a symbol (attack vector).