Add #cache_key to ImageTemplate
This method generates the cache key for each template.

Creates a checksum from all the files used to create the template.
(Simielar to what Rails does)
If any of them change, then the cache is cleared.

Looked into the possiblity of cache key collisions. TLDR it's possible
but probability wise, "winning powerball is more likely" to happen than
a collision.

My initial tests, this works really well. Seeing 500ms - 2 second
response time when uncached. When cached, it's under < 20ms.

Will need to update docs to mention LRU for redis (want to set so that
least used keys are expired first).