Big refactor. Swap to hmac digest auth. Remove redis/s3
Major simplification of the app. Removing S3 upload, going to
go straight to JPG and will let CDN cache it (or it doesn't really
matter since it should only get rendered twice).

Changing auth to hmac so that we can cache the URL's.

Removing redis, no caching needed. Simplifies the codebase a lot.